aptkit: always use noninteractive frontend
fis and gnome software can't manage debconf. gtk3widgets as the name states, are not adaptive and very ugly. if we were to allow debconf use an actual agent, it would have to be passthrough and implemented by the client (fis, gs)

for now always use noninteractive and call it a day. we don't have time for this right now

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>